remove ocp only logic and do the same things on k8s and ocp
