+++ api/v1/groupversion_info.go
@@ -26,7 +26,7 @@ import (
 
 var (
 	// GroupVersion is group version used to register these objects
-	GroupVersion = schema.GroupVersion{Group: "registry.apicur.io", Version: "v2"}
+	GroupVersion = schema.GroupVersion{Group: "registry.apicur.io", Version: "v1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
