+++ controllers/svc/client/client_monitoring.go
@@ -2,6 +2,7 @@ package client
 
 import (
 	ctx "context"
+
 	"github.com/Apicurio/apicurio-registry-operator/controllers/common"
 	"github.com/Apicurio/apicurio-registry-operator/controllers/loop/context"
 	monitoring "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
@@ -50,29 +51,6 @@ func (this *MonitoringClient) UpdateServiceMonitor(namespace common.Namespace, o
 	return this.client.ServiceMonitors(namespace.Str()).Update(ctx.TODO(), obj, v1.UpdateOptions{})
 }
 
-func (this *MonitoringClient) IsServiceMonitorRegistered() (bool, error) {
-	return resourceExists(this.discoveryClient, "monitoring.coreos.com/v1", "ServiceMonitor")
-}
-
 func (this *MonitoringClient) DeleteServiceMonitor(value *monitoring.ServiceMonitor, options *v1.DeleteOptions) error {
 	return this.client.ServiceMonitors(value.Namespace).Delete(ctx.TODO(), value.Name, *options)
-}
-
-// From k8sutil
-func resourceExists(dc discovery.DiscoveryInterface, apiGroupVersion, kind string) (bool, error) {
-
-	_, apiLists, err := dc.ServerGroupsAndResources()
-	if err != nil {
-		return false, err
-	}
-	for _, apiList := range apiLists {
-		if apiList.GroupVersion == apiGroupVersion {
-			for _, r := range apiList.APIResources {
-				if r.Kind == kind {
-					return true, nil
-				}
-			}
-		}
-	}
-	return false, nil
 }
